+++ adapters/PubMatic/PubMaticAdapter/PubMaticAdapter.swift
@@ -40,12 +40,14 @@ final class PubMaticAdapter: NSObject, RTBAdapter {
       let client = OpenWrapSDKClientFactory.createClient()
 
       // Sets COPPA compliance based on MobileAds configuration.
-      // - If either tag is true, treat as COPPA-compliant (true).
+      // - If either tag is true or ageRestrictedTreatment is set to child, treat as COPPA-compliant
+      //   (true).
       // - If either tag is false (and neither is true), treat as not COPPA-compliant (false).
       // - Otherwise, leave as nil.
       let isChild = MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment?.boolValue
       let isUnderAge = MobileAds.shared.requestConfiguration.tagForUnderAgeOfConsent?.boolValue
-      if isChild == true || isUnderAge == true {
+      let ageRestrictedTreatment = MobileAds.shared.requestConfiguration.ageRestrictedTreatment
+      if isChild == true || isUnderAge == true || ageRestrictedTreatment == .child {
         client.enableCOPPA(true)
       } else if isChild == false || isUnderAge == false {
         client.enableCOPPA(false)