+++ adapters/PubMatic/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## PubMatic iOS Mediation Adapter Changelog
 
+#### Next Version
+- Maps `ageRestrictedTreatment` to PubMatic's COPPA API.
+
 #### [Version 5.1.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pubmatic/PubMaticAdapter-5.1.0.0.zip)
 - Verified compatibility with PubMatic OpenWrap SDK version 5.1.0.